requirements.txt
# Flask Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
# SQLite is included with Python, no package needed

# Machine Learning
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
dill==0.3.7

# QR Code Generation
qrcode[pil]==7.4.2
Pillow==10.1.0

# Firebase Admin SDK
firebase-admin==6.4.0

# HTTP Requests
requests==2.31.0

# Production WSGI Server
gunicorn==21.2.0
gevent==23.9.1

# Environment Variables
python-dotenv==1.0.0

# Optional: For better performance
# uWSGI==2.0.23  # Alternative to gunicorn

